transfer

+++ run/create_dataset.sh
@@ -2,9 +2,9 @@
 
 # ************************************************** HELPSTRING **************************************************
 read -r -d '' HELP <<HELPSTRING
-Create dataset on Euler
-This script generates and submits batch jobs on Euler to generate a dataset. Creation is done as a job array while
-statistic aggregation is performed once all scripts have finished.
+Create dataset (on Euler or locally)
+This script generates and submits jobs on Euler or run them locally to generate a dataset. Creation is done as a job
+array on Euler and in parallel locally while statistic aggregation is performed once all scripts have finished.
 Usage:
     -c | --cfg                  : path to configuration file
     -n | --num_projects_per_job : number of projects in each array job
@@ -17,9 +17,14 @@ if [[ ${#ARGS[@]} == 0 ]]; then
   echo "Ivalid options. Use -h or --help for help."
   exit 1
 fi
+MACHINE=$1
+if [[ $MACHINE != "--euler" || $MACHINE != "--local" ]]; 
+    echo "invalid machine"; exit 1
+fi
+echo "Running on: $MACHINE"
 SHORT=,c:,s:,h
 LONG=cfg:,num_projects_per_job:,help
-OPTS=$(getopt -a -n dataset_creation --options $SHORT --longoption $LONG -- "$@")
+OPTS=$(getopt -a -n dataset_creation --options $SHORT --longoption $LONG -- "${@:2}")
 eval set -- "$OPTS"
 while :
 do
@@ -59,29 +64,57 @@ sub_cfgs=${sub_cfgs[@]:1}
 echo "Saved sub config files: $(dirname $CFG)"
 
 # *(2)* Submit preprocessing jobs
-echo "Submitting preprocessing job array ..."
-retvalue=($(sbatch /cluster/work/igp_psr/elwalt/pdm/rs-uncertainty/run/slurm/preprocess_projects.sh ${sub_cfgs[@]}))
-pp_job_array_id=${retvalue[-1]}
-echo "Done."
+if [[ $MACHINE == "euler" ]];
+then
+  echo "Submitting preprocessing job array ..."
+  retvalue=($(sbatch /cluster/work/igp_psr/elwalt/pdm/rs-uncertainty/run/slurm/preprocess_projects.sh ${sub_cfgs[@]}))
+  pp_job_array_id=${retvalue[-1]}
+  echo "Done."
+else #local
+  pp_pids=()
+  echo "Launching preprocessing processes ..."
+  for i in "${#CONFIG_FILES[@]}";
+  do
+    python /scratch/ewalt/pdm/rs-uncertainty/src/scripts/create_dataset.py --preprocess --cfg $CONFIG_FILES[$i] 1> dataset.$i.log 2> dataset.$i.log
+    pids+=($!)
+  done
+  echo "Done. Check logs: dataset.*.log"
+  for pid in "${pids[@]}"
+  do
+    wait $pid
+  done
+  echo "waiting on ${pids[@]} ..."
+  echo "Done."
+fi
 
 # *(3)* Submit aggregation job
-echo "Submitting aggregation job ..."
-retvalue=($(sbatch /cluster/work/igp_psr/elwalt/pdm/rs-uncertainty/run/slurm/aggregate_projects.sh $CFG $pp_job_array_id))
-agg_job_id=${retvalue[-1]}
-echo "Done."
-
-# *(4)* Delete sub config files
-echo "Deleting sub config files ..."
-glob=$(dirname $CFG)/*-*-*-*-*/
-rm -r $glob
-echo "Done."
+if [[ $MACHINE == "euler" ]];
+then
+  echo "Submitting aggregation job ..."
+  retvalue=($(sbatch /cluster/work/igp_psr/elwalt/pdm/rs-uncertainty/run/slurm/aggregate_projects.sh $CFG $pp_job_array_id))
+  agg_job_id=${retvalue[-1]}
+  echo "Done."
+else #local
+  echo "Launching aggregation process ..."
+  python /scratch/ewalt/pdm/rs-uncertainty/src/scripts/create_dataset.py --aggregate --cfg $CFG
+  pid=$!
+  wait $pid
+  echo "Done."
+  echo "Waiting on $pid ..."
+  echo "Done."
 
 # *(5)* Feedback
-echo "Job ids:"
-echo "- preprocessing job array : $pp_job_array_id"
-echo "- aggregation job array   : $agg_job_id"
+if [[ $MACHINE == "euler" ]];
+then
+  echo "Job ids:"
+  echo "- preprocessing job array : $pp_job_array_id"
+  echo "- aggregation job array   : $agg_job_id"
+fi
 
 # *(6)* kill jobs
-echo "[debug] Cancelling jobs ..."
-scancel $pp_job_array_id $agg_job_id
-echo "Done."
+if [[ $MACHINE == "euler" ]];
+then
+  echo "[debug] Cancelling jobs ..."
+  scancel $pp_job_array_id $agg_job_id
+  echo "Done."
+fi
